前后端开始对接

+++ src/main/java/com/lsq/community/service/UserService.java
@@ -1,11 +1,11 @@
 package com.lsq.community.service;
 
-import com.lsq.community.pojo.User;
+import com.lsq.community.po.User;
 
 /**
  * @author lishiqian
  * @Date:Created in 2018/3/8 18:05
  */
 public interface UserService {
-    User selectUserById(Long id);
+    User selectUserById(Integer id);
 }
